+++ tests/drf/__init__.py
@@ -1,0 +1,2 @@
+from .controller_with_output_template import ControllerWithOutputTemplate
+from .controller_with_serializer import ControllerWithSerializer