+++ tests/drf/controller_with_output_template.py
@@ -1,7 +1,7 @@
 from typing import Any
 from typing import Dict
 
-import winter
+import winter.drf
 
 
 @winter.controller
@@ -9,6 +9,6 @@ import winter
 class ControllerWithOutputTemplate:
 
     @winter.route_get('{?name}')
-    @winter.output_template('hello.txt')
+    @winter.drf.output_template('hello.txt')
     def greet(self, name: str) -> Dict[str, Any]:
         return {'name': name}