+++ tests/schema/test_response_schema.py
@@ -5,7 +5,7 @@ import pytest
 from drf_yasg import openapi
 
 import winter
-from winter.controller import get_component
+from winter.web.controller import get_component
 from winter.routing.routing import get_route
 from winter.schema.generation import build_response_schema
 from winter.schema.generation import build_responses_schemas