+++ winter/django/__init__.py
@@ -12,17 +12,17 @@ from django.conf.urls import url
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.request import Request
 
-from ..argument_resolver import arguments_resolver
-from ..controller import build_controller
-from ..controller import get_component
 from ..core import ComponentMethod
 from ..drf.auth import is_authentication_needed
 from ..exceptions.handlers import MethodExceptionsManager
-from ..output_processor import get_output_processor
 from ..routing.routing import Route
 from ..routing.routing import get_route
 from ..web import ResponseEntity
+from ..web.argument_resolver import arguments_resolver
+from ..web.controller import build_controller
+from ..web.controller import get_component
 from ..web.default_response_status import get_default_response_status
+from ..web.output_processor import get_output_processor
 from ..web.throttling import create_throttle_classes
 from ..web.urls import rewrite_uritemplate_with_regexps