+++ winter/routing/argument_resolvers/query_parameters_argument_resolver.py
@@ -5,8 +5,8 @@ from rest_framework.request import Request
 
 from ..routing import get_route
 from ... import type_utils
-from ...argument_resolver import ArgumentNotSupported
-from ...argument_resolver import ArgumentResolver
+from winter.web.argument_resolver import ArgumentNotSupported
+from winter.web.argument_resolver import ArgumentResolver
 from ...core import ArgumentDoesNotHaveDefault
 from ...core import ComponentMethodArgument
 from ...core.json import JSONDecodeException