+++ winter/web/__init__.py
@@ -11,3 +11,4 @@ from .response_header_serializer import response_headers_serializer
 from .response_status_annotation import response_status
 from .throttling import throttling
 from .urls import register_url_regexp
+from .controller import controller